% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_methods.R
\name{compare_methods}
\alias{compare_methods}
\title{compare_methods function}
\usage{
compare_methods(
  data,
  classes = NA,
  methods = c("CPM", "Linnorm"),
  psd = TRUE,
  show_plots = TRUE,
  perplexity = 30,
  plot_file_name = "tsne_plot.png",
  plot_colour_label = "Cell Types",
  plot_title = "tSNE embeddings"
)
}
\arguments{
\item{data}{scRNA count data - genes Vs cells (dim : g x c)}

\item{classes}{cell annotation i.e. the class of each of the cells in the data}

\item{methods}{vector of preprocessing methods to be compared}

\item{psd}{logical value to indicate if psd should be applied while comparing methods}

\item{show_plots}{logical value to indicate if plots should be shown}

\item{perplexity}{perplexity value while computing tSNE}

\item{plot_file_name}{name of the file to save the plot : defaults to "tsne_plot.png"}

\item{plot_colour_label}{title of the plot legend}

\item{plot_title}{title of the plot}
}
\value{
dataframe containing comparison of complexity scores
}
\description{
Returns a comparison of complexity scores of the data before and after
applying preprocessing methods and also after applying psdR.
Also plots the datasets in all above cases after applying tSNE
}
