% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexity.R
\name{complexity}
\alias{complexity}
\title{complexity function}
\usage{
complexity(data, classes = NA)
}
\arguments{
\item{data}{scRNA count data - genes Vs cells (dim : g x c)}

\item{classes}{cell annotation i.e. the class of each of the cells in data}
}
\value{
complexity score : higher score indicates lower complexity
}
\description{
Returns a complexity score of the data, based on Fischer's Discriminant Ratio.
Higher score indicates lower complexity
}
