% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_npsdr.R
\name{GD.npsdr.solver}
\alias{GD.npsdr.solver}
\title{Gradient descent based nonlinear sufficient dimension reduction machine via kernel trick}
\usage{
GD.npsdr.solver(
  x,
  y,
  H,
  h,
  lambda,
  delta,
  k = floor(length(y)/3),
  eps,
  max.iter,
  loss,
  a = NULL,
  c = NULL
)
}
\arguments{
\item{x}{data matrix}

\item{y}{either continuous or (+1,-1) typed binary response vector}

\item{H}{the number of slicing}

\item{h}{very small interval for calculating numerical derivatives for a given arbitrary loss function}

\item{lambda}{hyperparameter for the loss function}

\item{delta}{learning rate}

\item{k}{number of basis functions for a kernel trick}

\item{eps}{threshold for stopping iteration with respect to the magnitude of derivative}

\item{max.iter}{maximum iteration number for the optimization process}

\item{loss}{pre-specified loss functions are "logistic", svm","l2svm","lwpsvm", and user-defined loss function object also can be used formed by inside double quotation mark}

\item{a}{a hyperparameter for the LUM loss function}

\item{c}{a hyperparameter for the LUM loss function}
}
\value{
An estimated matrix B for a sufficient dimension reduction will be returned
}
\description{
A method for reducing a dimension of the data sufficiently using gradient descent optimization method
}
\examples{
\donttest{
set.seed(1)
n <- 200;
p <- 5;
H <- 20;
lambda <- 0.1
eps <- 1.0e-5
max.iter <- 10
h <- 1.0e-6; delta <- 5*1.0e-1
x <- matrix(rnorm(n*p, 0, 2), n, p)
err <- rnorm(n, 0, .2)
B <- matrix(0, p, 2)
B[1,1] <- 1; B[2,2] <- 1
x1 <- x \%*\% B[,1]
x2 <- x \%*\% B[,2]
fx <-  x1/(0.5 + (x2 + 1)^2)
y <- c(fx + err) # response
y.binary <- sign(y)
my.hinge <- function(m,...){
 rslt <- (1-m)*(as.numeric((1-m) > 0))
 return(rslt)
}
GD.npsdr.solver(x, y, H, h, lambda, delta, k=floor(length(y)/3), eps,
               max.iter, loss="my.logistic")
GD.npsdr.solver(x, y.binary, H=20, h, lambda, delta, k=floor(length(y)/3), eps,
               max.iter, loss="svm") }
}
\seealso{
\code{\link{pred.npsdr}}
}
\author{
Jungmin Shin, \email{jungminshin@korea.ac.kr}, Seungjun Shin, \email{sjshin@korea.ac.kr}
}
