% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_spectral_density.R
\name{GenerateExampleData}
\alias{GenerateExampleData}
\title{Produce example data set for demonstrating package functions}
\usage{
GenerateExampleData()
}
\value{
A data frame
}
\description{
Produce example data set for demonstrating package functions
}
