% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.psi.R
\name{get.psi}
\alias{get.psi}
\title{Kernel function generator}
\usage{
get.psi(x, y, k = floor(length(y)/3))
}
\arguments{
\item{x}{data matrix}

\item{y}{either continuous or (+1,-1) typed binary response vector}

\item{k}{number of basis functions for a kernel trick, floor(length(y)/3) is default}
}
\value{
generate the kernel function \eqn{\phi(\mathbf{x})}
\item{\code{w}}{the first d leading eigenvectors of the matrix \eqn{\mathbf{Q}\mathbf{K}\mathbf{Q}} in Li B. et. al., (2011)}
\item{\code{l}}{the first d leading eigenvectors of the matrix \eqn{\mathbf{Q}\mathbf{K}\mathbf{Q}}}
\item{\code{scaled.x}}{scaled \eqn{\mathbf{x}}}
\item{\code{bw}}{bandwidth of the kernel function}
\item{\code{k}}{parameter for the kernel, \code{floor(length(n)/3)} is applied as a default}
}
\description{
Generate a kernel function
}
\seealso{
\code{\link{npsdr}}, \code{\link{phix}}
}
\author{
Jungmin Shin, \email{jungminshin@korea.ac.kr}, Andreas Artemiou \email{artemiou@uol.ac.cy}, Seung Jun Shin, \email{sjshin@korea.ac.kr}
}
