% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gn.psdr.R
\name{Gn}
\alias{Gn}
\title{Determination of structured dimension}
\usage{
Gn(obj, x, rho, plot = FALSE, ...)
}
\arguments{
\item{obj}{The gdsdr object}

\item{x}{data matrix X}

\item{rho}{hyperparameter of CVBIC}

\item{plot}{if plot=TRUE, plot the estimated Gn values}

\item{...}{generic attributes}
}
\value{
Estimated Gn scores for a optimal structural dimension
}
\description{
Estimation of a structural dimensionality. Choose the k which maximizes a Gn value
}
\examples{
\donttest{
set.seed(1234)
n <- 200; p <- 5; H <- 10; lambda <- 0.001; eps <- 1.0e-6;
max.iter <- 10; h <- 1.0e-5; delta <- 2*1.0e-1;
init.theta <- rnorm(sd=1,n=p)
x <- matrix(rnorm(n*p, 0, 1), n, p)
err <- rnorm(n, 0, .2)
B <- matrix(0, p, 2)
B[1,1] <- 1; B[2,2] <- 1
x1 <- x \%*\% B[,1]
x2 <- x \%*\% B[,2]
fx <-  x1/(0.5 + (x2 + 1)^2)
y <- fx + err
obj <- GD.psdr.solver(x, y, init.theta, H,lambda, h, delta, eps, max.iter, loss="svm")
Gn(obj, x, rho=0.005, plot=TRUE) }
}
\seealso{
\code{\link{GD.psdr.solver}} \code{\link{plot.psdr}}
}
\author{
Jungmin Shin, \email{jungminshin@korea.ac.kr}, Seungjun Shin, \email{sjshin@korea.ac.kr}
}
