% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_npsdrx.R
\name{npsdrx}
\alias{npsdrx}
\title{Predict a response variable for the new explanatory variables}
\usage{
npsdrx(object, newdata, d = 2)
}
\arguments{
\item{object}{The object from function \code{npsdr}}

\item{newdata}{new data \eqn{\mathbf{X}}}

\item{d}{structural dimensionality. d=2 is default.}
}
\value{
the value of the estimated nonlinear mapping \eqn{\phi(\cdot)} is applied to
newdata \eqn{X} with dimension d is returned.
}
\description{
Returning new \eqn{\mathbf{X}} via the estimated nonlinear kernel.
}
\examples{
\donttest{
set.seed(1)
n <- 200;
p <- 5;
H <- 20;
lambda <- 0.1
eps <- 1.0e-5
max.iter <- 10
h <- 1.0e-6; delta <- 5*1.0e-1
x <- matrix(rnorm(n*p, 0, 2), n, p)
err <- rnorm(n, 0, .2)
B <- matrix(0, p, 2)
B[1,1] <- 1; B[2,2] <- 1
x1 <- x \%*\% B[,1]
x2 <- x \%*\% B[,2]
fx <-  x1/(0.5 + (x2 + 1)^2)
y <- c(fx + err) # response
y.binary <- sign(y)
new.x <- matrix(rnorm(n*p, 0, 2), n, p)
obj <- npsdr(x, y, H, h, lambda, delta, k=floor(length(y)/3), eps, max.iter, loss="svm")
npsdrx(object=obj, newdata=new.x, d = 2) }
}
\seealso{
\code{\link{npsdr}}
}
\author{
Jungmin Shin, \email{jungminshin@korea.ac.kr}, Andreas Artemiou \email{artemiou@uol.ac.cy}, Seung Jun Shin, \email{sjshin@korea.ac.kr}
}
