% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_phix.R
\name{phix}
\alias{phix}
\title{A function of yielding nonlinear mapping of X onto the reduced dimension}
\usage{
phix(value, obj, order = 2)
}
\arguments{
\item{value}{data matrix X}

\item{obj}{the object from function 'npsdr'}

\item{order}{the structural dimensionality}
}
\value{
Return the value of a nonlinear Mapping X to the reduced dimension of rank 2 via the kernel function
}
\description{
A function of yielding nonlinear mapping of X onto the reduced dimension
}
\examples{
\donttest{
set.seed(1)
n <- 200;
p <- 5;
H <- 20;
lambda <- 0.1
eps <- 1.0e-5
max.iter <- 10
h <- 1.0e-6; delta <- 5*1.0e-1
x <- matrix(rnorm(n*p, 0, 2), n, p)
err <- rnorm(n, 0, .2)
B <- matrix(0, p, 2)
B[1,1] <- 1; B[2,2] <- 1
x1 <- x \%*\% B[,1]
x2 <- x \%*\% B[,2]
fx <-  x1/(0.5 + (x2 + 1)^2)
y <- c(fx + err) # response
y.binary <- sign(y)
my.obj <- npsdr(x, y, H, h, lambda, delta, k=floor(length(y)/3), eps,
               max.iter, loss="svm")
x.nsvm <- phix(value=x, obj=my.obj, order=2)
}
}
\seealso{
\code{\link{npsdr}}
}
\author{
Jungmin Shin, \email{jungminshin@korea.ac.kr}, Seung Jun Shin, \email{sjshin@korea.ac.kr}
}
