% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_plot.psdr.R
\name{plot.psdr}
\alias{plot.psdr}
\title{scatter plot with the first predictor versus the response}
\usage{
\method{plot}{psdr}(obj, dim = 2, ...)
}
\arguments{
\item{obj}{The gdsdr list type output}

\item{dim}{decide number of dimensions that will be plotted, 2 is a default}

\item{...}{Generic compatibility}
}
\value{
scatter plot with transformed mth predictor versus response
}
\description{
produces a scatter plot of the transformed predictors versus response variable with lowess curve
}
\examples{
\donttest{
set.seed(1234)
n <- 200; p <- 5; H <- 10; lambda <- 0.001; eps <- 1.0e-6;
max.iter <- 10; h <- 1.0e-5; delta <- 2*1.0e-1;
init.theta <- rnorm(sd=1,n=p)
x <- matrix(rnorm(n*p, 0, 1), n, p)
err <- rnorm(n, 0, .2)
B <- matrix(0, p, 2)
B[1,1] <- 1; B[2,2] <- 1
x1 <- x \%*\% B[,1]
x2 <- x \%*\% B[,2]
fx <-  x1/(0.5 + (x2 + 1)^2)
y <- fx + err
y.binary <- sign(fx + err)
obj <- psdr(x, y, init.theta, H,lambda, h, delta, eps, max.iter, loss="svm")
plot(obj, dim=2) }
}
\seealso{
\code{\link{psdr}}, \code{\link{dimension}}
}
\author{
Jungmin Shin, Seung Jun Shin
}
