% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_pred.npsdr.R
\name{pred.npsdr}
\alias{pred.npsdr}
\title{Determination of structured dimension}
\usage{
pred.npsdr(obj, new.x, d = 2)
}
\arguments{
\item{obj}{The npsdr object}

\item{new.x}{new data X}

\item{d}{true dimension for data x. d=2 is default}
}
\value{
Estimated response variable y return
}
\description{
Estimation of a structural dimensionality. Choose the k which maximizes a Gn value
}
\examples{
\donttest{
set.seed(1)
n <- 200;
p <- 5;
H <- 20;
lambda <- 0.1
eps <- 1.0e-5
max.iter <- 10
h <- 1.0e-6; delta <- 5*1.0e-1
x <- matrix(rnorm(n*p, 0, 2), n, p)
err <- rnorm(n, 0, .2)
B <- matrix(0, p, 2)
B[1,1] <- 1; B[2,2] <- 1
x1 <- x \%*\% B[,1]
x2 <- x \%*\% B[,2]
fx <-  x1/(0.5 + (x2 + 1)^2)
y <- c(fx + err) # response
y.binary <- sign(y)
new.x <- matrix(rnorm(n*p, 0, 2), n, p)
obj <- GD.npsdr.solver(x, y, H, h, lambda, delta, k=floor(length(y)/3), eps, max.iter, loss="svm")
pred.npsdr(obj, new.x, d = 2) }
}
\seealso{
\code{\link{GD.npsdr.solver}}
}
\author{
Jungmin Shin, \email{jungminshin@korea.ac.kr}, Seungjun Shin, \email{sjshin@korea.ac.kr}
}
