% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_rtpsdr.R
\name{rtpsdr}
\alias{rtpsdr}
\title{Real time sufficient dimension reduction through principal least squares SVM}
\usage{
rtpsdr(A, r, n, Xbar, X, y, direction = "forward", H = NULL, lambda = NULL)
}
\arguments{
\item{A}{a temporal matrix produced by the previous estimation with original data}

\item{r}{an estimated parameters from the previous estimation with original data}

\item{n}{a number of sample size of the original data}

\item{Xbar}{a mean vector from the original data}

\item{X}{X in new data}

\item{y}{y in new data, y is continuous}

\item{direction}{A direction of change of the new data which is either "forward" or "backward". default is "forward"}

\item{H}{a number of slices. default is set to 10.}

\item{lambda}{hyperparameter for the loss function. default is set to 0.1.}
}
\value{
An updated psdr result will be returned in a real time fashion
}
\description{
Real time least squares SVM SDR method for a both regression and classification problem
}
\examples{
\donttest{
set.seed(1234)
n <- 300; n2 <- 500; p <- 5; H <- 10; lambda <- 0.1;
x <- matrix(rnorm(n*p, 0, 1), n, p)
x.new <- matrix(rnorm(n2*p, 0, 1), n2, p)
err <- rnorm(n, 0, .2); err.new <- rnorm(n2, 0, .2)
B <- matrix(0, p, 2)
B[1,1] <- 1; B[2,2] <- 1
x1 <- x \%*\% B[,1]
x2 <- x \%*\% B[,2]
x1.new <- x.new \%*\% B[,1]
x2.new <- x.new \%*\% B[,2]
fx <-  x1/(0.5 + (x2 + 1)^2)
fx.new <-  x1.new/(0.5 + (x2.new + 1)^2)
y <- fx + err;                     y.new <- fx.new + err.new
y.binary <- c(sign(fx + err));     y.binary.new <- c(sign(fx.new + err.new))
ls <- psdr(x, y, H=10, lambda=0.1, loss="ls")
wls <- psdr(x, y.binary, H=10, lambda=0.1, loss="wls")
rtpsdr(A=ls$A, r=ls$r, n=ls$N, Xbar=ls$Xbar, X=x.new, y=y.new, direction="forward", H, lambda)
rtpsdr(A=wls$A, r=wls$r, n=wls$N, Xbar=wls$Xbar, X=x.new, y=y.binary.new,
direction="forward", H, lambda)
}
}
\seealso{
\code{\link{psdr}}
}
\author{
Jungmin Shin, \email{jungminshin@korea.ac.kr}, Seung Jun Shin, \email{sjshin@korea.ac.kr}
}
