% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wisc.R
\docType{data}
\name{wisc}
\alias{wisc}
\title{Breast Cancer Wisconsin (Diagnostic) Data Set}
\format{
A data frame with 569 rows and 32 variables
}
\usage{
data(wisc)
}
\description{
Features are computed from a digitized image of a fine needle aspirate (FNA) of a breast mass.
They describe characteristics of the cell nuclei present in the image.
Ten real-valued features are computed for each cell nucleus:
}
\details{
\itemize{
\item ID number
\item Diagnosis (M = malignant, B = benign)
\item radius (mean of distances from center to points on the perimeter)
\item texture (standard deviation of gray-scale values)
\item perimeter
\item area
\item smoothness (local variation in radius lengths)
\item compactness (perimeter^2 / area - 1.0)
\item concavity (severity of concave portions of the contour)
\item concave points (number of concave portions of the contour)
\item symmetry
\item fractal dimension ("coastline approximation" - 1)
}
}
\keyword{datasets}
