% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_methods.R
\name{compare_methods}
\alias{compare_methods}
\title{compare_methods function}
\usage{
compare_methods(
  data,
  classes = NA,
  methods = c("CPM", "Linnorm"),
  psd = TRUE,
  show_plots = TRUE,
  plot_file_name = NA
)
}
\arguments{
\item{data}{scRNA count data - genes Vs cells (dim : g x c)}

\item{classes}{cell annotation i.e. the class of each of the cells in data}

\item{methods}{list of preprocessing methods to be compared}

\item{psd}{should psd be applied while comparing methods}

\item{show_plots}{should plots be shown}

\item{plot_file_name}{name of the file to save the plot}
}
\value{
comparison of complexity scores
}
\description{
Returns a comparison of complexity scores of the data before and after
applying preprocessing methods and also after applying psdR.
Also plots the datasets in all above cases after applying tSNE
}
