% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd.R
\name{psd}
\alias{psd}
\title{psd function}
\usage{
psd(data)
}
\arguments{
\item{data}{scRNA count data - cells Vs genes}
}
\value{
psd transformation of input data
}
\description{
Compute the Power Spectral Density (psd) transformation of the given scRNA count data
}
